* organização de dependencias
* adição do campo tipo de objeto

* separação de objetos na lista para NR

+++ src/main/java/br/com/caixa/sirfg/model/AmbienteEnum.java
@@ -1,20 +1,20 @@
 package br.com.caixa.sirfg.model;
 
 public enum AmbienteEnum {
-    DES(1, "DES"),
-    TQS(2, "TQS"),
-    HMP(3, "HMP"),
-    PRD(4, "PRD");
+    DES(1L, "DES"),
+    TQS(2L, "TQS"),
+    HMP(3L, "HMP"),
+    PRD(4L, "PRD");
 
-    private final int valorAmbiente;
+    private final Long valorAmbiente;
     private final String nomeAmbiente;
 
-    AmbienteEnum(int valorAmbiente, String nomeAmbiente) {
+    AmbienteEnum(Long valorAmbiente, String nomeAmbiente) {
         this.valorAmbiente = valorAmbiente;
         this.nomeAmbiente = nomeAmbiente;
     }
 
-    public int getValorAmbiente() {
+    public Long getValorAmbiente() {
         return valorAmbiente;
     }
